-- changed dir structure -- added requirements.txt, setup.py
