+++ requirements.txt
@@ -1,0 +1,2 @@
+matplotlib~=3.4.2
+setuptools~=57.0.0
