+++ setup.py
@@ -1,0 +1,14 @@
+from setuptools import setup
+
+setup(
+    name='Python_VLSI',
+    version='0.0.1',
+    packages=[''],
+    url='https://github.com/GeorgePap-719/Python_VLSI',
+    license='MIT',
+    author='geopapadopoulos',
+    author_email='george.719pap@gmail.com',
+    description='a simple setup script to manage dependencies'
+)
+
+# TODO transfer requirements.txt functionality here
