+++ vlsi/functions.py
@@ -4,6 +4,10 @@
 
 import os
 
+from vlsi.classes.Net import Net
+from vlsi.classes.Node import Node
+from vlsi.classes.Row import Row
+
 """
 folderName = "ibm01_mpl6_placed_and_nettetris_legalized"
 fileName = "ibm01"
@@ -15,200 +19,6 @@ os.chdir('C:\\Users\\root\\Desktop\\Python_Pandas\\docs\\ISPD\\{}'.format(
 folderName = "design"
 fileName = "design"
 os.chdir('C:\\Users\\root\\Desktop\\Python_Pandas\\docs\\{}'.format(folderName))
-
-""""    Classes    """
-
-
-class Point:
-
-    def __init__(self, x, y):
-        self.x = x
-        self.y = y
-
-    def __str__(self):
-        return " (" + str(self.x) + " , " + str(self.y) + ") "
-
-
-class Node:
-
-    def __init__(self, node_name, node_width, node_height, node_type,
-                 node_x=0, node_y=0):
-        self.node_name = node_name
-        self.node_width = node_width
-        self.node_height = node_height
-        self.node_type = node_type
-        self.node_x = node_x  # Lower Left Corner - x Coordinate
-        self.node_y = node_y  # Lower Left Corner - y Coordinate
-        self.lower_left_corner = Point(None, None)
-        self.lower_right_corner = Point(None, None)
-        self.upper_left_corner = Point(None, None)
-        self.upper_right_corner = Point(None, None)
-
-        self.node_nets = []  # net_names that this node are part of #TODO net objects
-        self.node_row = Row(None, None, None, None, None)  # row that this node is placed in
-
-    # update the Coordinates x & y
-    def set_x_y(self, node_x, node_y):
-        self.node_x = node_x
-        self.node_y = node_y
-
-    def set_row(self, row):
-        self.node_row = row
-
-    # update the list of nets
-    def append_net(self, net_name):
-        self.node_nets.append(str(net_name))
-
-    # calculate the coordinates of the 4-corners of the node
-    # Terminals are dots, they do not have corners
-    def set_points(self, x_min, x_max, y_min, y_max):
-        if self.node_type == "Terminal":
-            pass
-        else:
-            self.lower_left_corner = Point(x_min, y_min)
-            self.lower_right_corner = Point(x_max, y_min)
-            self.upper_left_corner = Point(x_min, y_max)
-            self.upper_right_corner = Point(x_max, y_max)
-
-    def display_node_corners(self):
-        print("\nNode name: " + str(self.node_name)
-              + "\nLower Left Corner: " + str(self.lower_left_corner)
-              + "\nLower Right Corner: " + str(self.lower_right_corner)
-              + "\nUpper Left Corner: " + str(self.upper_left_corner)
-              + "\nUpper Right Corner: " + str(self.upper_right_corner))
-
-    def display_node_row(self):
-        print("\nNode " + str(self.node_name)
-              + " is placed in row: " + str(self.node_row.row_name))
-
-    def display_node_nets(self):
-        print("\nNode " + str(self.node_name)
-              + " belongs to the net(s):  ")
-        for i in self.node_nets:
-            print(i, end=" ")
-
-    def __str__(self):
-        return (str(self.node_name) + " " + str(self.node_width) + " " +
-                str(self.node_height) + " " + str(self.node_type) + " " +
-                str(self.node_x) + " " + str(self.node_y))
-
-
-class Net:
-
-    def __init__(self, net_name, net_degree):
-        self.net_name = net_name
-        self.net_degree = net_degree
-        self.net_nodes = []  # list of nodes for the current net
-        self.net_rows = []  # list of rows that this net belongs to # TODO on parser()
-        self.x_min = None
-        self.x_max = None
-        self.y_min = None
-        self.y_max = None
-        self.wirelength = None
-        self.net_size = None
-
-    # update the list of nodes of this net
-    def append_node(self, node):
-        self.net_nodes.append(node)
-
-    def find_coordinates_of_net(self):
-        start = 0
-        for node in self.net_nodes:
-            start += 1
-
-            if start == 1 and node.node_type == "Non_Terminal":
-                self.x_min = node.lower_left_corner.x
-                self.x_max = node.lower_right_corner.x
-                self.y_min = node.lower_left_corner.y
-                self.y_max = node.upper_right_corner.y
-            elif start == 1 and node.node_type == "Terminal":
-                self.x_min = node.node_x
-                self.x_max = node.node_x
-                self.y_min = node.node_y
-                self.y_max = node.node_y
-            else:
-                if node.node_type == "Non_Terminal":
-                    if node.lower_left_corner.x < self.x_min:
-                        self.x_min = node.lower_left_corner.x
-                    if node.lower_right_corner.x > self.x_max:
-                        self.x_max = node.lower_right_corner.x
-                    if node.lower_left_corner.y < self.y_min:
-                        self.y_min = node.lower_left_corner.y
-                    if node.upper_right_corner.y > self.y_max:
-                        self.y_max = node.upper_right_corner.y
-                else:
-                    if node.node_x < self.x_min:
-                        self.x_min = node.node_x
-                    if node.node_x > self.x_max:
-                        self.x_max = node.node_x
-                    if node.node_y < self.y_min:
-                        self.y_min = node.node_y
-                    if node.node_y > self.y_max:
-                        self.y_max = node.node_y
-
-    def calculate_net_wirelength(self):
-        self.wirelength = (self.x_max - self.x_min) + (self.y_max - self.y_min)
-
-    def calculate_net_size(self):
-        self.net_size = (self.x_max - self.x_min) * (self.y_max - self.y_min)
-
-    def display_net(self):
-        print("\n" + str(self.net_name)
-              + " - netDegree =  " + str(self.net_degree))
-        print("Nodes of this net: ")
-        for node in self.net_nodes:
-            print(node.node_name, end=" ")
-
-    def display_net_size(self):
-        print(str(self.net_name) + " size = " + str(self.net_size))
-
-    def display_net_wirelength(self):
-        print(str(self.net_name) + " wirelength = " + str(self.wirelength))
-
-    # not displaying the nodes that are part of the net
-    def __str__(self):
-        return str(self.net_name) + " " + str(self.net_degree)
-
-
-class Row:
-
-    def __init__(self, row_name, y_min, y_max, x_min, x_max):
-        self.row_name = row_name
-        self.y_min = y_min
-        self.y_max = y_max
-        self.x_min = x_min
-        self.x_max = x_max
-        self.row_nodes = []  # list of nodes that are placed in this row
-        self.row_nets = []  # list of nets that are part of this row # TODO on parser()
-        self.lower_left_corner = Point(x_min, y_min)
-        self.lower_right_corner = Point(x_max, y_min)
-        self.upper_left_corner = Point(x_min, y_max)
-        self.upper_right_corner = Point(x_max, y_max)
-
-    # update the list of nodes of this row
-    def append_node(self, node):
-        if node.node_type == "Terminal":
-            pass
-        else:
-            self.row_nodes.append(node)
-
-    # update the list of nets of this row
-    def append_net(self, net):  # TODO on parser()
-        self.row_nets.append(net)
-
-    # display row name and nodes of this row
-    def display_row(self):
-        print("\n" + str(self.row_name))
-        print("Nodes of this row: ")
-        for i in self.row_nodes:
-            print(i, end=" ")
-
-    def __str__(self):
-        return (str(self.row_name) + " - y_min: "
-                + str(self.y_min) + " - y_max: "
-                + str(self.y_max) + " - x_min: "
-                + str(self.x_min) + " - x_max: " + str(self.x_max))
-
 
 """"    Functions   """
 
@@ -613,12 +423,12 @@ def parser():  # parsing the whole circuit
         a += 1
         i.display_net()
         i.find_coordinates_of_net()
-        i.calculate_net_wirelength()
+        i.calculate_net_wire_length()
         i.calculate_net_size()
 
         print("\n")
 
         i.display_net_size()
-        i.display_net_wirelength()
+        i.display_net_wire_length()
         if a == 15:
             break
