+++ vlsi/part1/Model.py
@@ -1,0 +1,8 @@
+# TODO
+class Model:
+
+    # TODO def __init__(self):
+
+    @staticmethod
+    def modeling():
+        stat = 1
